Add files via upload
The Company column is added to make forming a relationship between the tables easy in tableau (the only company is Citibike). The only difference between this version of the notebook and the last is it works on the complete csvs of Citibike data. Tableau has a hard time dealing with such large datasets however. For performance it will likely be better to use the provided sample datasets or take a random sampling from the large datasets.